fixing interaction with the headunit

+++ Pedal_Box/Pedal_Box_RP2040/src/buttonhandler.h
@@ -1,9 +1,9 @@
 class ButtonHandler
 {
     private:
-        unsigned int buttonID = 0, potOut = 0;
+        unsigned short buttonID = 0, potOut = 0;
     public:
-        short convertIDtoPot(String message){
+        unsigned short convertIDtoPot(String message){
             potOut = 0;
             buttonID = message.toInt();
             if (buttonID >= 0 && buttonID <= 9) //max radiout value
